move to common class
